+++ editor/Editor.hpp
@@ -13,7 +13,6 @@
 #include "EditorCamera.hpp"
 #include "ImGuiInputContext.hpp"
 #include "Project.hpp"
-#include "UI/States.hpp"
 
 #include <Game-Engine/Application.hpp>
 #include <Game-Engine/FrameGraph.hpp>
@@ -80,8 +79,6 @@ private:
     GE::ListScriptNamesFn m_listScriptNames;
     GE::ListScriptParametersFn m_listScriptParameters;
     GE::MakeScriptInstanceFn m_makeScriptInstance;
-
-    UI::States m_uiStates;
 
 public:
     Editor& operator=(const Editor&) = delete;
