--- editor/UI/EntityInspectorPanel.hpp
@@ -1,56 +1,0 @@
-/*
- * ---------------------------------------------------
- * EntityInspectorPanel.hpp
- *
- * Author: Thomas Choquet <semoir.dense-0h@icloud.com>
- * Date: 2024/11/04 15:03:54
- * ---------------------------------------------------
- */
-
-#ifndef ENTITYINSPECTORPANEL_HPP
-#define ENTITYINSPECTORPANEL_HPP
-
-#include <Game-Engine/Scene.hpp>
-#include <Game-Engine/Entity.hpp>
-#include <Game-Engine/ScriptLibraryManager.hpp>
-
-#include <functional>
-
-namespace GE_Editor
-{
-
-class EntityInspectorPanel
-{
-public:
-    EntityInspectorPanel() = delete;
-    EntityInspectorPanel(const EntityInspectorPanel&) = delete;
-    EntityInspectorPanel(EntityInspectorPanel&&) = delete;
-
-    EntityInspectorPanel(GE::Entity, GE::Scene*, GE::ListScriptNamesFn, GE::ListScriptParametersFn);
-
-    EntityInspectorPanel& onEntityDelete(std::function<void()>&& f) { return m_onEntityDelete = std::move(f), *this; }
-
-    void render();
-
-    ~EntityInspectorPanel() = default;
-
-private:
-    template<typename T> bool componentEditHeader(const char*);
-    template<typename T> void componentEditWidget();
-    void addComponentPopUp();
-
-    GE::Entity m_entity;
-    GE::Scene* m_scene;
-    GE::ListScriptNamesFn m_listScriptNames;
-    GE::ListScriptParametersFn m_listScriptParameters;
-
-    std::function<void()> m_onEntityDelete;
-
-public:
-    EntityInspectorPanel& operator=(const EntityInspectorPanel&) = delete;
-    EntityInspectorPanel& operator=(EntityInspectorPanel&&) = delete;
-};
-
-} // namespace GE_Editor
-
-#endif // ENTITYINSPECTORPANEL_HPP
